Code now works perfectly to get Wigner semicircle
Currently we are calculating the entire history of all walkers and writing it as a nested list in a Python file that can be used to do visualisations. First we need to make it more memory efficient by not maintaining the entire history in the RAM. Instead we only maintain two lists, the current step and the next step, and then after every time iteration we write to the file. That way we are not in need of exhorbitantly high amounts of RAM